the last commit of the decomposition plot
A value t was added to the function that controls the smoonthness of the
trend from the decompostion plot.
It takes values from 0 to 100.
Being 0, it is the default smoothness for frequencey bigger than 1.
Being 100, it is the most smoothness.
It could be useful to put a bar in the UI for  t.